Fix generate serverless project issue with absolute path.
Issue:
serverless cli will add prefix to any given path, absolute or not. This
will create project at incorrect location.

Solution:
When user use absolute path for the ouput location, we will generate the
project in a temporary folder in the current directory and mvoe to the
desired path when we finish.

We should submit a PR to the upstream serverless project to fix this
issue, instead of using this workaround