Fixed setup_sh feature in service_env (#486)
It appears that the setup_sh feature of the env decorator was broken.
We (@seantur and myself) implemented the functionality as described in the documentation
and added the tests seen below.

Additionally, we moved the setup.sh execution higher in the generated Dockerfile.
The reason for this was to allow the setup.sh script to add our private PyPI repository
to the image, which is required for the pip install -r /bento/requirements.txt line to
succeed for private packages. We would certainly welcome feedback on this solution
and discuss alternatives.