update build script
We are removing `.bin` directory, because when `pip install` it will cause `ERROR: Could not install packages due to an EnvironmentError: [Errno 2] No such file or directory` -> reason is Long install path.   Since we don't need the .bin directory in the dist version of web ui. We are going to remove it during the build script