JsonInput shouldn't check Content-Type for AWS Lambda functions (#904)
* Don't test Content-Type for aws lambda input

This addresses an issue with AWS Lambda where sometimes a `Content-Type` header is not being sent to the Lambda function even though it is set by the client. This resolves https://github.com/bentoml/BentoML/issues/893.

* Adapt wording of exception to reflect reality

* Update test to reflect changes in json handler

The json handler accepts any request to AWS Lambda as long as the body is valid json. 
This change reflects that in the tests.

* Remove trailing whitespaces

* Match the error output

* Auto format code

Co-authored-by: Chaoyu <paranoyang@gmail.com>