Fix zipkin module not found issue (#1532)
Fixed the "module not found" issue when the optional dependency zipkin not installed. 
The issue was due to the configuration injection tried to load all modules under `bentoml/tracing` package.
This PR also refactoring the tracing code path, fixed a number of issues related to opentracing/jaeger tracing, and updated documentation for tracing.