feat(tracing): add support for otlp exporter (#2918)
What does this PR address?

Mainly fixes #2917
However, it also addresses the following concerns:

Add opentelemetry-exporter-jaeger-thrift as an extra dependency in order to solve this issue
Update OpenTelemetry to 1.12 (no incompatibilities found on my side)
Log a warning message when sample_rate is not set for tracing.

Before submitting:




 Does the Pull Request follow Conventional Commits specification naming? Here are GitHub's
guide on how to create a pull request.
 Does the code follow BentoML's code style, both make format and make lint script have passed (instructions)?
 Did you read through contribution guidelines and follow development guidelines?
 Did your changes require updates to the documentation? Have you updated
those accordingly? Here are documentation guidelines and tips on writting docs.
 Did you write tests to cover your changes? -> No, I didn't found tests covering tracing. Open to do it in a subsequent PR though.

Who can help review?
It was mentioned in Slack that @bojiang worked on the tracing feature.