Modifyed bentos.py to construct HTTPServer using named arguments
I found an error in the way the serve function in bentos.py constructs the HTTPServer object. By not using named arguments and by omitting the timeout parameter, all the parameters below port  are interpreted as the wrong thing. Which on my system, caused a crash when the server tried to open a subprocess with invalid arguments.